+++ src/main/java/org/example/schedule/entity/Event.java
@@ -1,0 +1,5 @@
+package org.example.schedule.entity;
+
+public class Event {
+  String task;
+}
