+++ src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=Calendar
+spring.datasource.url=jdbc:mysql://localhost:3306/schedule
+spring.datasource.username=root
+spring.datasource.password=mysql
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+spring.sql.init.schema-locations=classpath:schedule.sql
+spring.sql.init.mode=always
