+++ src/test/java/org/example/schedule/CalendarApplicationTests.java
@@ -1,4 +1,4 @@
-package org.example.calendar;
+package org.example.schedule;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
